+++ src/Game/GameObjects/ConcreteWallGameObject.h
@@ -36,7 +36,7 @@ public:
 
     ConcreteWallGameObject(const WallGOType type, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void Render() const override;
-    virtual void UpdateFrame(const uint64_t delta) override;
+    virtual void UpdateFrame(const double delta) override;
 
 private:
     void RenderWallGo(const WallGOPosition position) const;
