+++ src/Game/GameObjects/Eagle.h
@@ -18,7 +18,7 @@ public:
 
     Eagle(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void Render() const override;
-    void UpdateFrame(const uint64_t delta) override;
+    void UpdateFrame(const double delta) override;
 
 private:
     std::array<std::shared_ptr<RenderEngine::Sprite>, 2> sprite;
