+++ src/Game/GameObjects/GameObjectInterface.h
@@ -6,7 +6,7 @@ class GameObjectInterface {
 public:
     GameObjectInterface(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void Render() const = 0;
-    virtual void UpdateFrame(const uint64_t delta) {};
+    virtual void UpdateFrame(const double delta) {};
     virtual ~GameObjectInterface();
 
 protected:
